:new: [efi-kern] Refactor efi-kern and run scripts to build bootx64.efi
Problem:
- We currently use make directly in the efi-kern directory which breaks
the unified cmake build system.
- We don't want to build in the source directory.
- Building bootx64.efi is not a simple enough process

Solution:
- Refactor cmake to replace the Makefile in efi-kern
- Refactor and move qos.img build/run scripts into root of project

Note: We assume that people are using docker to build all of the binaries.
However there is a disconnect between the build and run process. Running these
os images require root access, and we are not ready to simplify dockers process yet.
People who are eager may install qemu in their container and run docker in --privileged
mode to run the entire workflow from the container.